% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_ilr_coords.R
\name{append_ilr_coords}
\alias{append_ilr_coords}
\title{Add ILR coordinates to a data.frame containing composition variables}
\usage{
append_ilr_coords(dataf, comps, psi)
}
\arguments{
\item{dataf}{data.frame containing composition variables}

\item{comps}{character vector of composition variable names in dataf}

\item{psi}{ilrBase passed to \code{compositions::ilr()}}
}
\description{
Add ILR coordinates to a data.frame containing composition variables
}
