% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_strictly_positive_vals.R
\name{check_strictly_positive_vals}
\alias{check_strictly_positive_vals}
\title{Check if compositional variable are strictly greater than 0}
\usage{
check_strictly_positive_vals(dataf, comps, tol = 1e-06)
}
\arguments{
\item{dataf}{data.frame containing composition variables}

\item{comps}{character vector of composition variable names in dataf}

\item{tol}{a numeric value that compositional values are expected to be greater or equal than. 1e-6 is deafult}
}
\value{
If any compositional values are found to be strictly less than \code{tol} and erro is thrown. 
Returns \code{TRUE} invisibly otherwise.
}
\description{
Check if compositional variable are strictly greater than 0
}
