% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lm_quantities.R
\name{extract_lm_quantities}
\alias{extract_lm_quantities}
\title{Extract critical quantities from a lm object (for confidence interval calculations)}
\usage{
extract_lm_quantities(lm_X, alpha = 0.05)
}
\arguments{
\item{lm_X}{a lm object}

\item{alpha}{level of significance. Defaults to 0.05.}
}
\value{

}
\description{
Extract critical quantities from a lm object (for confidence interval calculations)
}
\examples{
data(fat_data)
lm_fat <- lm(fat ~ sl, data = fat_data)
extract_lm_quantities(lm_fat)
}
