% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_pred.R
\name{get_mean_pred}
\alias{get_mean_pred}
\title{Title}
\usage{
get_mean_pred(lm_obj, newdata, alpha = 0.05)
}
\arguments{
\item{lm_obj}{}

\item{newdata}{}

\item{alpha}{}
}
\value{

}
\description{
Title
}
