% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plus_minus_changes.R
\name{get_plus_minus_changes}
\alias{get_plus_minus_changes}
\title{Get predictions from compositional ilr multiple linear regression model}
\usage{
get_plus_minus_changes(dataf, y, comps, covars = NULL, deltas = c(0, 10,
  20)/(24 * 60), comparisons = c("one-v-one", "one-v-all", "prop-realloc")[1],
  alpha = 0.05, verbose = FALSE)
}
\arguments{
\item{dataf}{A \code{data.frame} containing data}

\item{y}{Name (as string) of outcome in \code{dataf}}

\item{comps}{Character vector of names of compositions in \code{dataf}}

\item{covars}{Optional. Character vector of covariates names  (non-comp variables) in \code{dataf}. Defaults to NULL.}

\item{deltas}{Optional. Changes in compositions to be computed pairwise. Defaults to 0, 10 and 20 minutes as a proportion of minutes in a day.}

\item{comparisons}{Currently three choices: "one-v-one" (default), "one-v-all" or "prop-realloc". Currently proportional re-allocaiton isn't properly implemented.}

\item{alpha}{Optional. Level of significance. Defaults to 0.05.}

\item{verbose}{Optional. Whether the function provides extra information upon return. Defaults to FALSE.}
}
\description{
Provided the data (containing outcome, composiitional compoents and covariates), fit a ilr multiple linear regression model and provide predictions from reallocating compositional values pairwise amunsnst the components model.
}
\examples{
get_plus_minus_changes(
  dataf=fat_data,
  y="fat",
  comps=c("sl","sb","lpa","mvpa"),
  covars=c("sibs","parents","ed"),
  deltas=seq(-60,60,by=5)/(24*60),
  comparisons="one-v-one",
  alpha=0.05,
  verbose=FALSE
)

}
\author{
Ty Stanford <tystan@gmail.com>
}
