% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pred_bounds.R
\name{get_pred_bounds}
\alias{get_pred_bounds}
\title{Title}
\usage{
get_pred_bounds(y0_star, crit_val, se_y0_star, bound = 0)
}
\arguments{
\item{y0_star}{}

\item{crit_val}{}

\item{se_y0_star}{}

\item{bound}{}
}
\value{

}
\description{
Title
}
