% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_realloc_nms.R
\name{get_realloc_nms}
\alias{get_realloc_nms}
\title{Title}
\usage{
get_realloc_nms(comps, comparisons, poss_comps)
}
\arguments{
\item{comps}{}

\item{comparisons}{}

\item{poss_comps}{}
}
\value{

}
\description{
Title
}
