% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x0_star.R
\name{get_x0_star}
\alias{get_x0_star}
\title{Title}
\usage{
get_x0_star(dmX, n_preds, ilr_names, ilr_delta, ilr_means)
}
\arguments{
\item{dmX}{}

\item{n_preds}{}

\item{ilr_names}{}

\item{ilr_delta}{}

\item{ilr_means}{}
}
\value{

}
\description{
Title
}
