% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_deltacomp_obj.R
\name{is_deltacomp_obj}
\alias{is_deltacomp_obj}
\title{Is object that is returned from \code{pred_delta_comps()}?}
\usage{
is_deltacomp_obj(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Boolean TRUE or FALSE
}
\description{
Is object that is returned from \code{pred_delta_comps()}?
}
