% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_lm_mod.R
\name{is_lm_mod}
\alias{is_lm_mod}
\title{Is object that is returned from \code{lm()}?}
\usage{
is_lm_mod(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Boolean TRUE or FALSE
}
\description{
Is object that is returned from \code{lm()}?
}
