% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_null_or_na.R
\name{is_null_or_na}
\alias{is_null_or_na}
\title{Catch NULL, empty and objects containing NAs}
\usage{
is_null_or_na(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Boolean. If object is NULL, empty or contains NA then TRUE returned. FALSE otherwise.
}
\description{
Catch NULL, empty and objects containing NAs
}
