% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_delta_comps.R
\name{predict_delta_comps}
\alias{predict_delta_comps}
\title{Get predictions from compositional ilr multiple linear regression model}
\usage{
predict_delta_comps(
  dataf,
  y,
  comps,
  covars = NULL,
  deltas = c(0, 10, 20)/(24 * 60),
  comparisons = c("prop-realloc", "one-v-one")[1],
  alpha = 0.05
)
}
\arguments{
\item{dataf}{A \code{data.frame} containing data}

\item{y}{Name (as string) of outcome in \code{dataf}}

\item{comps}{Character vector of names of compositions in \code{dataf}}

\item{covars}{Optional. Character vector of covariates names  (non-comp variables) in \code{dataf}. Defaults to NULL.}

\item{deltas}{A vector of time-component changes (as proportions of compositions , i.e., values between -1 and 1). Optional. Changes in compositions to be computed pairwise. Defaults to 0, 10 and 20 minutes as a proportion of the 1440 minutes in a day.}

\item{comparisons}{Currently three choices: "one-v-one" or "prop-realloc" (default). Currently "one-v-all"  isn't properly implemented.}

\item{alpha}{Optional. Level of significance. Defaults to 0.05.}
}
\description{
Provided the data (containing outcome, composiitional compoents and covariates), fit a ilr multiple linear regression model and provide predictions from reallocating compositional values pairwise amunsnst the components model.
}
\examples{
predict_delta_comps(
  dataf=fat_data,
  y = "fat",
  comps = c("sl", "sb", "lpa", "mvpa"),
  covars = c("sibs", "parents", "ed"),
  deltas = seq(-60, 60, by = 5) / (24 * 60),
  comparisons = "one-v-one",
  alpha = 0.05
)

}
\author{
Ty Stanford <tystan@gmail.com>
}
