% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ilr_trans.R
\name{print_ilr_trans}
\alias{print_ilr_trans}
\title{Print the ilr transformation of provided composition parts to console}
\usage{
print_ilr_trans(comps)
}
\arguments{
\item{comps}{a character vector of compositional parts}
}
\value{
a character vector of representing the ilr transformation of the \code{comps}
is returned invisibly as the function's purpose is simply to 
print to the R console
}
\description{
Print the ilr transformation of provided composition parts to console
}
