% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_exist.R
\name{cols_exist}
\alias{cols_exist}
\title{Check whether columns exist in a data.frame}
\usage{
cols_exist(dataf, cols)
}
\arguments{
\item{dataf}{a data.frame}

\item{cols}{character vector of columns to be checked in \code{dataf}}
}
\value{
An error if all \code{cols} not present in \code{dataf}. 
Returns \code{TRUE} invisibly otherwise.
}
\description{
Check whether columns exist in a data.frame
}
