% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lm.R
\name{fit_lm}
\alias{fit_lm}
\title{fit linear model based on input data.frame}
\usage{
fit_lm(y_str, X, verbose = TRUE)
}
\arguments{
\item{y_str}{a string representation of the column in \code{X} that is the outcome}

\item{X}{a data.frame or matrix that contains the predictor and outcome variables}

\item{verbose}{if \code{TRUE} (default), a model summary will be printed to the console}
}
\value{
A \code{lm} object where the \code{y_str} column has been regressed against the remaining
columns of \code{X} (with an intercept term as well).
}
\description{
fit linear model based on input data.frame
}
