% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_delta_comps.R
\name{predict_delta_comps}
\alias{predict_delta_comps}
\title{Get predictions from compositional ilr multiple linear regression model}
\usage{
predict_delta_comps(
  dataf,
  y,
  comps,
  covars = NULL,
  deltas = c(0, 10, 20)/(24 * 60),
  comparisons = c("prop-realloc", "one-v-one")[1],
  alpha = 0.05
)
}
\arguments{
\item{dataf}{A \code{data.frame} containing data}

\item{y}{Name (as string/character vector of length 1) of outcome variable in \code{dataf}}

\item{comps}{Character vector of names of compositions in \code{dataf}. See details for more information.}

\item{covars}{Optional. Character vector of covariates names  (non-comp variables) in \code{dataf}. Defaults to NULL.}

\item{deltas}{A vector of time-component changes (as proportions of compositions , i.e., values between -1 and 1). Optional. 
Changes in compositions to be computed pairwise. Defaults to 0, 10 and 20 minutes as a proportion of the 1440 minutes 
in a day (i.e., approximately \code{0.000}, \code{0.007} and \code{0.014}).}

\item{comparisons}{Currently two choices:  \code{"one-v-one"} or  \code{"prop-realloc"} (default). Please see details for explanation of these methods.}

\item{alpha}{Optional. Level of significance. Defaults to 0.05.}
}
\description{
Provided the data (containing outcome, compositional components and covariates), fit a ilr multiple linear regression model and provide predictions from reallocating compositional values pairwise amunsnst the components model.
}
\details{
Values in the \code{comps} columns must be strictly greater than zero. These compositional values are NOT assumed to be constrained to (0, 1) 
values as the function normalises the compositions row-wise to sum to 1 in part of it's processing of the dataset before analysis.

Please see the \code{deltacomp} package \code{README.md} file for examples and explanation of the \code{comparisons = "prop-realloc"} and \code{comparisons = "one-v-one"} options. 

Note from version 0.1.0 to current version, \code{comparisons == "one-v-all"} is depreciated, \code{comparisons == "prop-realloc"} is probably the alternative you are after.
}
\examples{
predict_delta_comps(
  dataf = fat_data,
  y = "fat",
  comps = c("sl", "sb", "lpa", "mvpa"),
  covars = c("sibs", "parents", "ed"),
  deltas = seq(-60, 60, by = 5) / (24 * 60),
  comparisons = "one-v-one",
  alpha = 0.05
)

predict_delta_comps(
  dataf = fat_data,
  y = "fat",
  comps = c("sl", "sb", "lpa", "mvpa"),
  covars = NULL,
  deltas = seq(-60, 60, by = 5) / (24 * 60),
  comparisons = "one-v-one",
  alpha = 0.05
)

}
\author{
Ty Stanford <tystan@gmail.com>
}
